refactor: 优化测试用例，测试 lib 而不是 src
